requirements.txt
attrs==23.1.0
certifi==2023.5.7
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
jsonlines==3.1.0
outcome==1.2.0
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.1
trio-websocket==0.10.3
urllib3==2.0.3
wsproto==1.2.0